add support for SRv6 uN behavior
The SRv6 "micro segment" (SRv6 uSID or uSID for short) instruction is
a straightforward extension of the SRv6 Network Programming model [1].

This patch implements the SRv6 uN behavior which is variant of the endpoint
behavior (End). The SRv6 uN behavior takes the uSID Block and the uSID,
which are defined as follows:

 - uSID Block: A Block of uSID's. It can be any IPv6 prefix available
               to the provider. The uSID Block length is expressed in bits
               and it must be divisible by 8 (byte aligned). For instance,
               valid uSID Block lengths  are: 8, 16, 32, etc;

 - uSID:       The "micro segment" ID. The uSID length is expressed in bits
               and must be divisible by 8 (byte aligned). For instance,
               valid uSID lengths  are: 8, 16, 32, etc.

In this SRv6 uN behavior implementation: 1) if the uSID Block length is not
supplied, then it is set to 32 bits; 2) if the uSID length is not supplied,
then it is set to 16 bits.

[1] https://tools.ietf.org/html/draft-filsfils-spring-net-pgm-extension-srv6-usid-04

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>
Signed-off-by: Paolo Lungaroni <paolo.lungaroni@cnit.it>