add sysctl support for SRv6 micro segment behaviors
If CONFIG_SYSCTL is set, then default uSID Block and uSID lengths can be
read/written using the sysctl tool at run-time. The uSID Block and uSID
lengths are available under:

 - uSID Block length: (sysctl) net.ipv6.seg6_local.usid_block_len or
                      (procfs) /proc/sys/net/ipv6/seg6_local/usid_block_len;

 - uSID length:       (sysctl) net.ipv6.seg6_local.usid_len or
                      (procfs) /proc/sys/net/ipv6/seg6_local/usid_len.

If CONFIG_SYSCTL is not set, the uSID Block and uSID lengths are
respectively set to 32 bits and 16 bits.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>
Signed-off-by: Paolo Lungaroni <paolo.lungaroni@cnit.it>